Use long keys for simulator events

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/Synthetic.java
@@ -15,7 +15,7 @@
  */
 package com.github.benmanes.caffeine.cache.simulator;
 
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import com.github.benmanes.caffeine.cache.simulator.BasicSettings.SyntheticSettings.HotspotSettings;
 import com.github.benmanes.caffeine.cache.simulator.BasicSettings.SyntheticSettings.UniformSettings;
@@ -38,7 +38,7 @@ public final class Synthetic {
   private Synthetic() {}
 
   /** Returns a sequence of events based on the setting's distribution. */
-  public static IntStream generate(BasicSettings settings) {
+  public static LongStream generate(BasicSettings settings) {
     int items = settings.synthetic().events();
     switch (settings.synthetic().distribution().toLowerCase()) {
       case "counter":
@@ -70,7 +70,7 @@ public final class Synthetic {
    * @param start the number that the counter starts from
    * @param items the number of items in the distribution
    */
-  public static IntStream counter(int start, int items) {
+  public static LongStream counter(int start, int items) {
     return generate(new CounterGenerator(start), items);
   }
 
@@ -81,7 +81,7 @@ public final class Synthetic {
    * @param mean mean arrival rate of gamma (a half life of 1/gamma)
    * @param items the number of items in the distribution
    */
-  public static IntStream exponential(double mean, int items) {
+  public static LongStream exponential(double mean, int items) {
     return generate(new ExponentialGenerator(mean), items);
   }
 
@@ -98,7 +98,7 @@ public final class Synthetic {
    * @param hotOpnFraction percentage of operations accessing the hot set
    * @param items the number of items in the distribution
    */
-  public static IntStream hotspot(int lowerBound, int upperBound,
+  public static LongStream hotspot(int lowerBound, int upperBound,
       double hotsetFraction, double hotOpnFraction, int items) {
     return generate(new HotspotIntegerGenerator(lowerBound,
         upperBound, hotsetFraction, hotOpnFraction), items);
@@ -112,7 +112,7 @@ public final class Synthetic {
    *
    * @param items the number of items in the distribution
    */
-  public static IntStream scrambledZipfian(int items) {
+  public static LongStream scrambledZipfian(int items) {
     return generate(new ScrambledZipfianGenerator(items), items);
   }
 
@@ -122,7 +122,7 @@ public final class Synthetic {
    *
    * @param items the number of items in the distribution
    */
-  public static IntStream skewedZipfianLatest(int items) {
+  public static LongStream skewedZipfianLatest(int items) {
     return generate(new SkewedLatestGenerator(new CounterGenerator(items)), items);
   }
 
@@ -132,7 +132,7 @@ public final class Synthetic {
    *
    * @param items the number of items in the distribution
    */
-  public static IntStream zipfian(int items) {
+  public static LongStream zipfian(int items) {
     return generate(new ZipfianGenerator(items), items);
   }
 
@@ -145,12 +145,12 @@ public final class Synthetic {
    * @param items the number of items in the distribution
    * @return a stream of cache events
    */
-  public static IntStream uniform(int lowerBound, int upperBound, int items) {
+  public static LongStream uniform(int lowerBound, int upperBound, int items) {
     return generate(new UniformIntegerGenerator(lowerBound, upperBound), items);
   }
 
   /** Returns a sequence of items constructed by the generator. */
-  private static IntStream generate(IntegerGenerator generator, int count) {
-    return IntStream.range(0, count).map(ignored -> generator.nextInt());
+  private static LongStream generate(IntegerGenerator generator, long count) {
+    return LongStream.range(0, count).map(ignored -> generator.nextInt());
   }
 }
